+++ ceri-common/src/main/java/ceri/common/math/ReverseFunction.java
@@ -20,6 +20,7 @@ public class ReverseFunction {
 		for (int i = 0; i <= steps; i++) {
 			double x = steps == 0 ? x0 : x0 + (i * (x1 - x0) / steps);
 			double y = fn.applyAsDouble(x);
+			//noinspection SuspiciousNameCombination
 			b.add(y, x);
 		}
 		return b.build();
@@ -89,10 +90,10 @@ public class ReverseFunction {
 		if (floor == null && ceiling == null) return Double.NaN;
 		if (floor == null) return ceiling.getValue();
 		if (ceiling == null) return floor.getValue();
-		double y0 = floor.getKey().doubleValue();
-		double y1 = ceiling.getKey().doubleValue();
-		double x0 = floor.getValue().doubleValue();
-		double x1 = ceiling.getValue().doubleValue();
+		double y0 = floor.getKey();
+		double y1 = ceiling.getKey();
+		double x0 = floor.getValue();
+		double x1 = ceiling.getValue();
 		if (y0 == y1) return x0;
 		return x0 + ((y - y0) * (x1 - x0) / (y1 - y0));
 	}
